refactor: CarController, 메인 -> RacingController에 통합
