--- src/main/java/racingcar/ApplicationMain.java
@@ -1,11 +1,0 @@
-package racingcar;
-
-import racingcar.controller.RacingController;
-
-public class ApplicationMain {
-	public static void main(String[] args) {
-		RacingController racingController = new RacingController();
-		racingController.createCarsByUserInput();
-		racingController.run();
-	}
-}
