+++ src/test/java/racingcar/controller/RacingControllerTest.java
@@ -8,25 +8,25 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class CarControllerTest {
+public class RacingControllerTest {
 
 	@Test
 	@DisplayName("자동차 이름 콤마 기준 분리, 공백 제거")
 	void splitCarNamesTest() {
-		CarController carController = new CarController();
+		RacingController racingController = new RacingController();
 		List<String> expected = Arrays.asList("범고래", "소주캉", "고래");
 		String userInputCarNames = "  범고래 , 소주캉,   고래";
 
-		assertThat(carController.splitCarNames(userInputCarNames)).isEqualTo(expected);
+		assertThat(racingController.splitCarNames(userInputCarNames)).isEqualTo(expected);
 	}
 
 	@Test
 	@DisplayName("3대의 자동차 목록 생성 테스트")
 	void createCarsTest() {
-		CarController carController = new CarController();
+		RacingController racingController = new RacingController();
 		String userInputCarNames = "범고래,소주캉,고래";
-		carController.createCars(userInputCarNames);
+		racingController.createCars(userInputCarNames);
 
-		assertThat(carController.getCars().size()).isEqualTo(3);
+		assertThat(racingController.getCars().size()).isEqualTo(3);
 	}
 }
